+++ src/plugin-hooks/useExpanded.js
@@ -5,19 +5,18 @@ import { expandRows } from '../utils'
 import {
   useGetLatest,
   actions,
-  functionalUpdate,
   useMountedLayoutEffect,
   makePropGetter,
 } from '../publicUtils'
 
 // Actions
-actions.toggleExpanded = 'toggleExpanded'
-actions.toggleAllExpanded = 'toggleAllExpanded'
-actions.setExpanded = 'setExpanded'
 actions.resetExpanded = 'resetExpanded'
+actions.toggleRowExpanded = 'toggleRowExpanded'
+actions.toggleAllRowsExpanded = 'toggleAllRowsExpanded'
 
 export const useExpanded = hooks => {
-  hooks.getExpandedToggleProps = [defaultGetExpandedToggleProps]
+  hooks.getToggleAllRowsExpandedProps = [defaultGetToggleAllRowsExpandedProps]
+  hooks.getToggleRowExpandedProps = [defaultGetToggleRowExpandedProps]
   hooks.stateReducers.push(reducer)
   hooks.useInstance.push(useInstance)
   hooks.prepareRow.push(prepareRow)
@@ -25,17 +24,29 @@ export const useExpanded = hooks => {
 
 useExpanded.pluginName = 'useExpanded'
 
-const defaultGetExpandedToggleProps = (props, { row }) => [
+const defaultGetToggleAllRowsExpandedProps = (props, { instance }) => [
   props,
   {
     onClick: e => {
-      e.persist()
-      row.toggleExpanded()
+      instance.toggleAllRowsExpanded()
     },
     style: {
       cursor: 'pointer',
     },
-    title: 'Toggle Expanded',
+    title: 'Toggle All Rows Expanded',
+  },
+]
+
+const defaultGetToggleRowExpandedProps = (props, { row }) => [
+  props,
+  {
+    onClick: () => {
+      row.toggleRowExpanded()
+    },
+    style: {
+      cursor: 'pointer',
+    },
+    title: 'Toggle Row Expanded',
   },
 ]
 
@@ -55,14 +66,32 @@ function reducer(state, action, previousState, instance) {
     }
   }
 
-  if (action.type === actions.setExpanded) {
+  if (action.type === actions.toggleAllRowsExpanded) {
+    const { value } = action
+    const { isAllRowsExpanded, rowsById } = instance
+
+    const expandAll = typeof value !== 'undefined' ? value : !isAllRowsExpanded
+
+    if (expandAll) {
+      const expanded = {}
+
+      Object.keys(rowsById).forEach(rowId => {
+        expanded[rowId] = true
+      })
+
+      return {
+        ...state,
+        expanded,
+      }
+    }
+
     return {
       ...state,
-      expanded: functionalUpdate(action.expanded, state.expanded),
+      expanded: {},
     }
   }
 
-  if (action.type === actions.toggleExpanded) {
+  if (action.type === actions.toggleRowExpanded) {
     const { id, value: setExpanded } = action
     const exists = state.expanded[id]
 
@@ -93,15 +122,27 @@ function useInstance(instance) {
   const {
     data,
     rows,
+    rowsById,
     manualExpandedKey = 'expanded',
     paginateExpandedRows = true,
     expandSubRows = true,
     autoResetExpanded = true,
+    getHooks,
     state: { expanded },
     dispatch,
   } = instance
 
   const getAutoResetExpanded = useGetLatest(autoResetExpanded)
+
+  let isAllRowsExpanded = Boolean(
+    Object.keys(rowsById).length && Object.keys(expanded).length
+  )
+
+  if (isAllRowsExpanded) {
+    if (Object.keys(rowsById).some(id => !expanded[id])) {
+      isAllRowsExpanded = false
+    }
+  }
 
   // Bypass any effects from firing when this changes
   useMountedLayoutEffect(() => {
@@ -110,10 +151,15 @@ function useInstance(instance) {
     }
   }, [dispatch, data])
 
-  const toggleExpanded = React.useCallback(
+  const toggleRowExpanded = React.useCallback(
     (id, value) => {
-      dispatch({ type: actions.toggleExpanded, id, value })
+      dispatch({ type: actions.toggleRowExpanded, id, value })
     },
+    [dispatch]
+  )
+
+  const toggleAllRowsExpanded = React.useCallback(
+    value => dispatch({ type: actions.toggleAllRowsExpanded, value }),
     [dispatch]
   )
 
@@ -129,20 +175,30 @@ function useInstance(instance) {
     expanded,
   ])
 
+  const getInstance = useGetLatest(instance)
+
+  const getToggleAllRowsExpandedProps = makePropGetter(
+    getHooks().getToggleAllRowsExpandedProps,
+    { instance: getInstance() }
+  )
+
   Object.assign(instance, {
     preExpandedRows: rows,
     expandedRows,
     rows: expandedRows,
-    toggleExpanded,
     expandedDepth,
+    isAllRowsExpanded,
+    toggleRowExpanded,
+    toggleAllRowsExpanded,
+    getToggleAllRowsExpandedProps,
   })
 }
 
 function prepareRow(row, { instance: { getHooks }, instance }) {
-  row.toggleExpanded = set => instance.toggleExpanded(row.id, set)
+  row.toggleRowExpanded = set => instance.toggleRowExpanded(row.id, set)
 
-  row.getExpandedToggleProps = makePropGetter(
-    getHooks().getExpandedToggleProps,
+  row.getToggleRowExpandedProps = makePropGetter(
+    getHooks().getToggleRowExpandedProps,
     {
       instance,
       row,